[symex] Move the per-thread state and SSA naming into goto-symex/state/
Fifth of seven commits partitioning src/goto-symex/
(docs/roadmap/goto-symex-partition-plan.md §2). goto_symex_state carries what a
single thread's execution holds; renaming is the L0/L1/L2 naming layer beneath
it. Neither knows about scheduling.

Repoints the two SYMEX-HARNESS-TARGET pins in regression/esbmc/symex_ssa_00 —
drift_check.py resolves them by path. The recorded sha256 is unchanged: it is
taken over the definition text, which a move does not touch.
