+++ src/goto-symex/execution_state.h
@@ -4,8 +4,8 @@
 #include <algorithm>
 #include <deque>
 #include <goto-symex/goto_symex.h>
-#include <goto-symex/goto_symex_state.h>
-#include <goto-symex/renaming.h>
+#include <goto-symex/state/goto_symex_state.h>
+#include <goto-symex/state/renaming.h>
 #include <goto-symex/equation/symex_target.h>
 
 #include <list>
