+++ src/goto-symex/state/renaming.h
