+++ src/goto-symex/symex_main.cpp
@@ -2,7 +2,7 @@
 #include <optional>
 #include <goto-symex/execution_state.h>
 #include <goto-symex/goto_symex.h>
-#include <goto-symex/goto_symex_state.h>
+#include <goto-symex/state/goto_symex_state.h>
 #include <goto-symex/reachability_tree.h>
 #include <goto-symex/equation/symex_target_equation.h>
