+++ unit/goto-symex/renaming.test.cpp
@@ -32,7 +32,7 @@
 #include <string>
 
 #include <goto-symex/reachability_tree.h>
-#include <goto-symex/renaming.h>
+#include <goto-symex/state/renaming.h>
 #include <irep2/irep2_expr.h>
 #include <util/lang/c_types.h>
 #include <util/symtab/namespace.h>
